Move companion window subcomponents into it's own namespace
