+++ app/components/companion_windows/content_list_component.html.erb
@@ -6,5 +6,5 @@
     data-controller="content-list"
     data-action="thumbnails-found@window->content-list#drawThumbnails"
     role="tablist">
-    <%= render ContentListItemComponent.with_collection(viewer.purl_object.downloadable_files, viewer:) %>
+    <%= render CompanionWindows::ContentListItemComponent.with_collection(viewer.purl_object.downloadable_files, viewer:) %>
 </ul>
