+++ app/components/companion_windows/content_list_item_component.html.erb
