+++ app/components/companion_windows/rights_component.html.erb
