--- app/components/content_list_component.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-class ContentListComponent < ViewComponent::Base
-  def initialize(viewer:)
-    @viewer = viewer
-  end
-
-  attr_reader :viewer
-
-  def heading
-    t('content_list_heading', scope: viewer.i18n_path)
-  end
-end
